Update and organize Initialize class methods
Rearranged methods in Initialize class, redefined and repurposed the `generateConfig` method to return an object output and introduced `writeConfig` method to conduct the writing operation. The modifications were made to improve functional clarity and explicitness. Adaptations within `lib/index.js` were executed to reflect these changes.
