+++ lib/Initialize.js
@@ -1,8 +1,7 @@
 const fs = require('fs');
 const path = require('path');
-const {defaultWpInstallLanguage, defaultWpConfig, defaultWpInstallFolder} = require("./constants");
-const {getLastWpVersion} = require("./utils/wpConfig");
-const {getUserLocale} = require("./utils");
+const {defaultWpConfig, defaultWpInstallFolder, pkgFileName, defaultWpDatabaseType} = require("./constants.js");
+const {getLastWp, getUserLocale} = require("./utils/wpConfig.js");
 
 /**
  * Initialize class for generating WordPress configuration file.
@@ -15,49 +14,50 @@ class Initialize {
    */
   constructor () {
     this.wpFolder = process.cwd();
-    this.outputPath = path.join(this.wpFolder, 'wp-package.json');
+    this.outputPath = path.join(this.wpFolder, pkgFileName);
   }
 
-  hasConfig = () => fs.existsSync(this.outputPath);
+  hasConfig = () => !! fs.existsSync( path.join( this.outputPath, pkgFileName ) );
 
   /**
    * Generates the configuration file for WordPress.
    *
-   * @return {void}
+   * @return {WPMMconfig} The configuration object.
    */
-  generateConfig () {
+  generateConfig() {
 
     // check if the output path exists
-    if (this.hasConfig) {
-      console.error(`👍 The wp-package.json file ${this.outputPath} already exists.`);
+    if (this.hasConfig()) {
+      console.log(`👍 The configuration file ${this.outputPath} already exists.`);
       return;
     }
 
     const name = this.wpFolder.split(path.sep).pop() || defaultWpInstallFolder;
-    const language = getUserLocale() || defaultWpInstallLanguage;
-    const version = getLastWpVersion();
-    const backupFolder = path.join(this.wpFolder, 'backups');
+    const lastWp = getLastWp();
+    const userLocale = getUserLocale();
 
-    const result = {
+    return {
       name,
       wordpress: {
-        version,
-        language,
+        version: lastWp.version,
+        language: userLocale,
         config: defaultWpConfig
       },
       database: {
-        type: "mysql",
-        backupFolder: backupFolder
+        type: defaultWpDatabaseType,
+        backupFolder: path.join(this.wpFolder, 'backups')
       },
       themes: [],
       plugins: [],
       postInstall: []
     };
+  }
 
+  writeConfig = (result) => {
     // write the config to the output path
     fs.writeFileSync(this.outputPath, JSON.stringify(result, null, 2));
     console.log(`🆗 Wordpress configuration file created. Configuration saved to ${this.outputPath}`);
-  }
+  };
 }
 
 module.exports = Initialize;
