+++ lib/index.js
@@ -1,5 +1,5 @@
 #!/usr/bin/env node
-const { getConfig, printTimePassed, getInfo } = require('./utils/index.js');
+const { getConfig, printTimePassed, getInfo } = require('./utils/data.js');
 const Installer = require('./installer');
 const Dump = require('./dump.js');
 const Database = require('./database.js');
@@ -97,7 +97,8 @@ const actions = {
   /** Initialize the WordPress installation. */
   init: () => {
     const initializer = new Initialize();
-    initializer.generateConfig();
+    const result = initializer.generateConfig();
+    initializer.writeConfig(result);
   },
 
   /** Upload a database by executing SQL queries from a specified file. */
@@ -105,6 +106,25 @@ const actions = {
     const db = new Database(config);
     db.uploadDatabase(file || config.database.filename).then(() => {
       console.log('🚀 Database uploaded successfully.');
+    });
+  },
+
+  /** Install WordPress */
+  install: () => {
+    const installer = new Installer(config, paths);
+
+    installer.run().then(() => {
+      console.log('🚀 WordPress installed successfully.');
+    });
+  },
+
+  /** Update WordPress packages */
+  update: () => {
+    const updater = new Updater(config);
+    updater.run(argv).then(() => {
+      console.log('🚀 WordPress updated successfully.');
+    }).catch(() => {
+      console.log('🔴 WordPress update failed.');
     });
   },
 
@@ -123,25 +143,6 @@ const actions = {
     actions["dump-database"]() && console.log('🚀 WP Database dumped successfully.');
     actions.dump() && console.log('🚀 All data dumped successfully.');
   },
-
-  /** Install WordPress */
-  install: () => {
-    const installer = new Installer(config);
-
-    installer.run().then(() => {
-      console.log('🚀 WordPress installed successfully.');
-    });
-  },
-
-  /** Update WordPress packages */
-  update: () => {
-    const updater = new Updater(config);
-    updater.run(argv).then(() => {
-      console.log('🚀 WordPress updated successfully.');
-    }).catch(() => {
-      console.log('🔴 WordPress update failed.');
-    });
-  }
 };
 
 const action = Object.keys(argv).find(key => argv[key] === true);
